+++ ImageFeed/Helpers/URLSession.swift
@@ -37,8 +37,8 @@ extension URLSession {
                 }
             } else if let error = error {
                 fulfillCompletion(.failure(NetworkError.urlRequestError(error)))
-            } else { fulfillCompletion(.failure(NetworkError.urlSessionError))
-                
+            } else {
+                fulfillCompletion(.failure(NetworkError.urlSessionError))
             }
         })
         task.resume()
